+++ packages/sdk/jest.config.js
@@ -58,13 +58,6 @@ module.exports = {
   // A path to a module which exports an async function that is triggered once after all test suites
   // globalTeardown: undefined,
 
-  // A set of global variables that need to be available in all test environments
-  globals: {
-    "ts-jest": {
-      "tsconfig": path.join(__dirname, "tsconfig.json")
-    }
-  },
-
   // The maximum amount of workers used to run your tests. Can be specified as % or a number. E.g. maxWorkers: 10% will use 10% of your CPU amount + 1 as the maximum worker number. maxWorkers: 2 will use a maximum of 2 workers.
   // maxWorkers: "50%",
 
@@ -172,7 +165,7 @@ module.exports = {
 
   // A map from regular expressions to paths to transformers
   transform: {
-    "^.+\\.ts$": "<rootDir>/../../node_modules/ts-jest"
+    "^.+\\.ts$": ["@swc/jest"]
   },
 
   // An array of regexp pattern strings that are matched against all source file paths, matched files will skip transformation