+++ packages/sdk/src/base.ts
@@ -1,5 +1,6 @@
 import { unzlibSync, zlibSync } from "fflate";
 
+import { isLyraInitialized, LyraState, transformPcmToLyra, transformLyraToPcm } from "./lyra";
 import {
   ConnectError,
   createDataChannelData,
@@ -16,10 +17,12 @@ import {
   trace,
 } from "./utils";
 import {
+  AudioCodecType,
   Callbacks,
   ConnectionOptions,
   JSONType,
   DataChannelConfiguration,
+  RTCEncodedAudioFrame,
   SignalingConnectMessage,
   SignalingMessage,
   SignalingNotifyMessage,
@@ -181,6 +184,24 @@ export default class ConnectionBase {
    * E2EE インスタンス
    */
   protected e2ee: SoraE2EE | null;
+  /**
+   * mid と AudioCodecType の対応づけを保持するマップ
+   *
+   * Lyra などのカスタム音声コーデック使用時に RTCRtpReceiver をどのコーデックでデコードすべきかを
+   * 判別するために使われる
+   *
+   * カスタム音声コーデックが有効になっていない場合には空のままとなる
+   */
+  private midToAudioCodecType: Map<string, AudioCodecType> = new Map();
+  /**
+   * Lyra などのカスタムコーデックでの変換用に生成した TransformStream を
+   * WebRTC 切断時に綺麗に終了させるための AbortController
+   */
+  private encodedTransformAbortController?: AbortController;
+  /**
+   * Lyra インスタンス
+   */
+  private lyra?: LyraState;
 
   constructor(
     signalingUrlCandidates: string | string[],
@@ -252,6 +273,13 @@ export default class ConnectionBase {
     this.signalingOfferMessageDataChannels = {};
     this.connectedSignalingUrl = "";
     this.contactSignalingUrl = "";
+    if (isLyraInitialized()) {
+      if (options.e2ee === true) {
+        console.warn("Currently, it is not possible to use E2EE and Lyra at the same time (Lyra is disabled).");
+      } else {
+        this.lyra = new LyraState();
+      }
+    }
   }
 
   /**
@@ -901,6 +929,11 @@ export default class ConnectionBase {
       }
       this.callbacks.disconnect(event);
     }
+
+    if (this.encodedTransformAbortController !== undefined) {
+      this.encodedTransformAbortController.abort();
+      this.encodedTransformAbortController = undefined;
+    }
   }
 
   /**
@@ -1171,7 +1204,7 @@ export default class ConnectionBase {
    */
   protected async connectPeerConnection(message: SignalingOfferMessage): Promise<void> {
     let config = Object.assign({}, message.config);
-    if (this.e2ee) {
+    if (this.e2ee || this.lyra) {
       // @ts-ignore https://w3c.github.io/webrtc-encoded-transform/#specification
       config = Object.assign({ encodedInsertableStreams: true }, config);
     }
@@ -1228,7 +1261,9 @@ export default class ConnectionBase {
     if (!this.pc) {
       return;
     }
-    const sessionDescription = new RTCSessionDescription({ type: "offer", sdp: message.sdp });
+
+    const sdp = this.processOfferSdp(message.sdp);
+    const sessionDescription = new RTCSessionDescription({ type: "offer", sdp });
     await this.pc.setRemoteDescription(sessionDescription);
     this.writePeerConnectionTimelineLog("set-remote-description", sessionDescription);
     return;
@@ -1281,16 +1316,172 @@ export default class ConnectionBase {
         // setRemoteDescription 後でないと active が反映されないのでもう一度呼ぶ
         await this.setSenderParameters(transceiver, this.encodings);
         const sessionDescription = await this.pc.createAnswer();
+        // TODO(sile): 動作確認
+        if (sessionDescription.sdp !== undefined) {
+          sessionDescription.sdp = this.processAnswerSdpForLocal(sessionDescription.sdp);
+        }
         await this.pc.setLocalDescription(sessionDescription);
         this.trace("TRANSCEIVER SENDER GET_PARAMETERS", transceiver.sender.getParameters());
         return;
       }
     }
     const sessionDescription = await this.pc.createAnswer();
+    if (sessionDescription.sdp !== undefined) {
+      sessionDescription.sdp = this.processAnswerSdpForLocal(sessionDescription.sdp);
+    }
     this.writePeerConnectionTimelineLog("create-answer", sessionDescription);
     await this.pc.setLocalDescription(sessionDescription);
     this.writePeerConnectionTimelineLog("set-local-description", sessionDescription);
     return;
+  }
+
+  /**
+   * カスタムコーデック対応用に offer SDP を処理するメソッド
+   *
+   * @param sdp offer SDP
+   * @returns 処理後の SDP
+   */
+  private processOfferSdp(sdp: string): string {
+    if (this.lyra === undefined || !sdp.includes("109 lyra/")) {
+      return sdp;
+    }
+
+    // mid と音声コーデックの対応を保存する
+    this.midToAudioCodecType.clear();
+    for (const media of sdp.split(/^m=/m).slice(1)) {
+      if (!media.startsWith("audio")) {
+        continue;
+      }
+
+      const mid = /a=mid:(.*)/.exec(media);
+      if (mid) {
+        const codecType = media.includes("109 lyra/") ? "LYRA" : "OPUS";
+        this.midToAudioCodecType.set(mid[1], codecType);
+      }
+    }
+
+    return this.lyra.processOfferSdp(sdp);
+  }
+
+  /**
+   * カスタムコーデック用に answer SDP を処理するメソッド
+   *
+   * 処理後の SDP は setLocalDescription() メソッドに渡される
+   *
+   * @param answer SDP
+   * @returns 処理後の SDP
+   */
+  private processAnswerSdpForLocal(sdp: string): string {
+    if (this.lyra === undefined) {
+      return sdp;
+    }
+
+    return this.lyra.processAnswerSdpForLocal(sdp);
+  }
+
+  /**
+   * カスタムコーデック用に answer SDP を処理するメソッド
+   *
+   * 処理後の SDP は Sora に送信される
+   *
+   * @param answer SDP
+   * @returns 処理後の SDP
+   */
+  private processAnswerSdpForSora(sdp: string): string {
+    if (this.lyra === undefined) {
+      return sdp;
+    }
+
+    return this.lyra.processAnswerSdpForSora(sdp);
+  }
+
+  /**
+   * 必要なら sender をカスタムコーデックを使ってエンコードする
+   *
+   * @param sender 対象となる RTCRtpSender インスタンス
+   */
+  protected async setupSenderTransformForCustomCodec(sender: RTCRtpSender): Promise<void> {
+    // TODO(sile): E2EE との併用を考慮する
+    if (this.lyra === undefined || sender.track === null) {
+      return;
+    }
+
+    if (this.encodedTransformAbortController === undefined) {
+      this.encodedTransformAbortController = new AbortController();
+    }
+    const signal = this.encodedTransformAbortController.signal;
+
+    // TODO(sile): WebRTC Encoded Transform の型が提供されるようになったら ignore を外す
+    // @ts-ignore
+    // eslint-disable-next-line
+    const senderStreams = sender.createEncodedStreams() as TransformStream;
+    const isLyraCodec = sender.track.kind === "audio" && this.options.audioCodecType === "LYRA";
+    if (isLyraCodec) {
+      const lyraEncoder = await this.lyra.createEncoder();
+      const transformStream = new TransformStream({
+        transform: (data: RTCEncodedAudioFrame, controller) => transformPcmToLyra(lyraEncoder, data, controller),
+      });
+      senderStreams.readable
+        .pipeThrough(transformStream, { signal })
+        .pipeTo(senderStreams.writable)
+        .catch((e) => {
+          lyraEncoder.destroy();
+          if (!signal.aborted) {
+            console.warn(e);
+          }
+        });
+    } else {
+      senderStreams.readable.pipeTo(senderStreams.writable, { signal }).catch((e) => {
+        if (!signal.aborted) {
+          console.warn(e);
+        }
+      });
+    }
+  }
+
+  /**
+   * 必要なら receiver をカスタムコーデックを使ってデコードする
+   *
+   * @param mid コーデックの判別に使う mid
+   * @param receiver 対象となる RTCRtpReceiver インスタンス
+   */
+  protected async setupReceiverTransformForCustomCodec(mid: string | null, receiver: RTCRtpReceiver): Promise<void> {
+    // TODO(sile): E2EE との併用を考慮する
+    if (this.lyra === undefined) {
+      return;
+    }
+
+    if (this.encodedTransformAbortController === undefined) {
+      this.encodedTransformAbortController = new AbortController();
+    }
+    const signal = this.encodedTransformAbortController.signal;
+
+    // TODO(sile): WebRTC Encoded Transform の型が提供されるようになったら ignore を外す
+    // @ts-ignore
+    // eslint-disable-next-line
+    const receiverStreams = receiver.createEncodedStreams() as TransformStream;
+    const codecType = this.midToAudioCodecType.get(mid || "");
+    if (codecType == "LYRA") {
+      const lyraDecoder = await this.lyra.createDecoder();
+      const transformStream = new TransformStream({
+        transform: (data: RTCEncodedAudioFrame, controller) => transformLyraToPcm(lyraDecoder, data, controller),
+      });
+      receiverStreams.readable
+        .pipeThrough(transformStream, { signal })
+        .pipeTo(receiverStreams.writable)
+        .catch((e) => {
+          lyraDecoder.destroy();
+          if (!signal.aborted) {
+            console.warn(e);
+          }
+        });
+    } else {
+      receiverStreams.readable.pipeTo(receiverStreams.writable, { signal }).catch((e) => {
+        if (!signal.aborted) {
+          console.warn(e);
+        }
+      });
+    }
   }
 
   /**
@@ -1299,7 +1490,8 @@ export default class ConnectionBase {
   protected sendAnswer(): void {
     if (this.pc && this.ws && this.pc.localDescription) {
       this.trace("ANSWER SDP", this.pc.localDescription.sdp);
-      const message = { type: "answer", sdp: this.pc.localDescription.sdp };
+      const sdp = this.processAnswerSdpForSora(this.pc.localDescription.sdp);
+      const message = { type: "answer", sdp };
       this.ws.send(JSON.stringify(message));
       this.writeWebSocketSignalingLog("send-answer", message);
     }